refactor(jobs): split unified sqlite codecs
