+++ crates/axon-jobs/src/unified.rs
@@ -5,6 +5,7 @@ use uuid::Uuid;
 
 use crate::boundary::{JobStore, Result};
 use crate::state_machine::validate_transition;
+use crate::unified_codec::*;
 
 #[derive(Debug, Clone)]
 pub struct SqliteUnifiedJobStore {
@@ -336,196 +337,6 @@ impl JobStore for SqliteUnifiedJobStore {
         }
         .into())
     }
-}
-
-fn descriptor(summary: &JobSummary) -> JobDescriptor {
-    JobDescriptor {
-        job_id: summary.job_id,
-        kind: summary.kind,
-        status: summary.status,
-        poll: PollDescriptor {
-            status_url: format!("/v1/jobs/{job_id}", job_id = summary.job_id.0),
-            events_url: Some(format!(
-                "/v1/jobs/{job_id}/events",
-                job_id = summary.job_id.0
-            )),
-            suggested_interval_ms: 1000,
-        },
-        created_at: summary.created_at.clone(),
-        updated_at: summary.updated_at.clone(),
-    }
-}
-
-async fn find_by_idempotency_key(
-    pool: &SqlitePool,
-    idempotency_key: &str,
-) -> Result<Option<JobSummary>> {
-    sqlx::query("SELECT * FROM jobs WHERE idempotency_key = ?")
-        .bind(idempotency_key)
-        .fetch_optional(pool)
-        .await
-        .map_err(sql_error)?
-        .map(row_to_summary)
-        .transpose()
-}
-
-async fn ensure_job_pool(pool: &SqlitePool, job_id: JobId) -> Result<()> {
-    let found = sqlx::query_scalar::<_, i64>("SELECT COUNT(*) FROM jobs WHERE job_id = ?")
-        .bind(job_id.0.to_string())
-        .fetch_one(pool)
-        .await
-        .map_err(sql_error)?;
-    if found == 0 {
-        return Err(missing_job(job_id));
-    }
-    Ok(())
-}
-
-async fn ensure_job(tx: &mut sqlx::Transaction<'_, sqlx::Sqlite>, job_id: JobId) -> Result<()> {
-    let found = sqlx::query_scalar::<_, i64>("SELECT COUNT(*) FROM jobs WHERE job_id = ?")
-        .bind(job_id.0.to_string())
-        .fetch_one(&mut **tx)
-        .await
-        .map_err(sql_error)?;
-    if found == 0 {
-        return Err(missing_job(job_id));
-    }
-    Ok(())
-}
-
-fn row_to_summary(row: sqlx::sqlite::SqliteRow) -> Result<JobSummary> {
-    Ok(JobSummary {
-        job_id: JobId::new(parse_uuid(row.get::<String, _>("job_id"))?),
-        kind: parse_enum(row.get::<String, _>("kind"))?,
-        intent: parse_optional_enum(row.get::<Option<String>, _>("intent"))?,
-        status: parse_enum(row.get::<String, _>("status"))?,
-        phase: parse_enum(row.get::<String, _>("phase"))?,
-        created_at: Timestamp(row.get("created_at")),
-        updated_at: Timestamp(row.get("updated_at")),
-        started_at: row.get::<Option<String>, _>("started_at").map(Timestamp),
-        finished_at: row.get::<Option<String>, _>("finished_at").map(Timestamp),
-        source_id: row.get::<Option<String>, _>("source_id").map(SourceId::new),
-        watch_id: row.get::<Option<String>, _>("watch_id").map(WatchId::new),
-        parent_job_id: row
-            .get::<Option<String>, _>("parent_job_id")
-            .map(parse_uuid)
-            .transpose()?
-            .map(JobId::new),
-        root_job_id: row
-            .get::<Option<String>, _>("root_job_id")
-            .map(parse_uuid)
-            .transpose()?
-            .map(JobId::new),
-        attempt: row.get::<i64, _>("attempt") as u32,
-        priority: parse_enum(row.get::<String, _>("priority"))?,
-        counts: from_optional_json(row.get::<Option<String>, _>("counts_json"))?,
-        current: from_optional_json(row.get::<Option<String>, _>("current_json"))?,
-        heartbeat: from_optional_json(row.get::<Option<String>, _>("heartbeat_json"))?,
-        last_error: from_optional_json(row.get::<Option<String>, _>("last_error_json"))?,
-        warnings: from_json(row.get::<String, _>("warnings_json"))?,
-    })
-}
-
-fn row_to_event(row: sqlx::sqlite::SqliteRow) -> Result<JobEvent> {
-    Ok(JobEvent {
-        event_id: row.get("event_id"),
-        sequence: row.get::<i64, _>("sequence") as u64,
-        job_id: JobId::new(parse_uuid(row.get::<String, _>("job_id"))?),
-        attempt: row.get::<i64, _>("attempt") as u32,
-        stage_id: row
-            .get::<Option<String>, _>("stage_id")
-            .map(parse_uuid)
-            .transpose()?
-            .map(StageId::new),
-        phase: parse_enum(row.get::<String, _>("phase"))?,
-        status: parse_enum(row.get::<String, _>("status"))?,
-        severity: parse_enum(row.get::<String, _>("severity"))?,
-        visibility: parse_enum(row.get::<String, _>("visibility"))?,
-        message: row.get("message"),
-        timestamp: Timestamp(row.get("timestamp")),
-        details: from_json(row.get::<String, _>("details_json"))?,
-    })
-}
-
-fn enum_name<T: serde::Serialize>(value: T) -> Result<String> {
-    serde_json::to_value(value)
-        .ok()
-        .and_then(|value| value.as_str().map(ToOwned::to_owned))
-        .ok_or_else(|| ApiError::new("job.enum_invalid", ErrorStage::Planning, "invalid enum"))
-}
-
-fn parse_enum<T: serde::de::DeserializeOwned>(value: String) -> Result<T> {
-    serde_json::from_value(serde_json::Value::String(value)).map_err(json_error)
-}
-
-fn parse_optional_enum<T: serde::de::DeserializeOwned>(value: Option<String>) -> Result<Option<T>> {
-    value.map(parse_enum).transpose()
-}
-
-fn to_json<T: serde::Serialize>(value: &T) -> Result<String> {
-    serde_json::to_string(value).map_err(json_error)
-}
-
-fn optional_to_json<T: serde::Serialize>(value: &Option<T>) -> Result<Option<String>> {
-    value.as_ref().map(to_json).transpose()
-}
-
-fn from_json<T: serde::de::DeserializeOwned>(value: String) -> Result<T> {
-    serde_json::from_str(&value).map_err(json_error)
-}
-
-fn from_optional_json<T: serde::de::DeserializeOwned>(value: Option<String>) -> Result<Option<T>> {
-    value.map(from_json).transpose()
-}
-
-fn parse_uuid(value: String) -> Result<Uuid> {
-    Uuid::parse_str(&value).map_err(|error| {
-        ApiError::new(
-            "job.uuid_invalid",
-            ErrorStage::Retrieving,
-            format!("invalid job uuid: {error}"),
-        )
-    })
-}
-
-fn now_timestamp() -> Timestamp {
-    Timestamp::from(chrono::Utc::now())
-}
-
-fn is_terminal(status: LifecycleStatus) -> bool {
-    matches!(
-        status,
-        LifecycleStatus::Completed
-            | LifecycleStatus::CompletedDegraded
-            | LifecycleStatus::Failed
-            | LifecycleStatus::Canceled
-            | LifecycleStatus::Expired
-            | LifecycleStatus::Skipped
-    )
-}
-
-fn escape_sql(value: &str) -> String {
-    value.replace('\'', "''")
-}
-
-fn missing_job(job_id: JobId) -> ApiError {
-    ApiError::new(
-        "job.not_found",
-        ErrorStage::Retrieving,
-        format!("job {} not found", job_id.0),
-    )
-}
-
-fn sql_error(error: sqlx::Error) -> ApiError {
-    ApiError::new(
-        "job.sqlite_error",
-        ErrorStage::Publishing,
-        error.to_string(),
-    )
-}
-
-fn json_error(error: serde_json::Error) -> ApiError {
-    ApiError::new("job.json_error", ErrorStage::Publishing, error.to_string())
 }
 
 #[cfg(test)]
